requirements.txt
flask
flask-cors
opencv-python-headless
mediapipe
numpy
Pillow
